feat: multi-notification system, daily goals/milestones, settings redesign, tooltips, vacuum DB
- Multi-notification system with three presets (VideoInfo, GeneralMessage, GoalMet)
- NotificationManager with stacked QPointer list, repositioning, and cleanup
- Right-click pauses notification and resumes same instance through QPointer chain
- Daily video/time goals with fractional support (2 decimals) and milestones
- Progress persistence via maininfo table with QDate-based day-change handling
- Silent catch-up on restart to prevent re-firing
- Settings dialog redesigned with sidebar navigation (QListWidget + QStackedWidget)
- Notifications page with preset toggles/durations and milestone step config
- Global tooltip system (TooltipEventFilter) with enable/disable and configurable delay
- Tooltips added to buttons and interactive controls across all dialogs
- Parent-chain ancestor walk fixes child-widget tooltip suppression
- Vacuum DB button with reclaimed space reporting and error logging
- Code review fixes: PRAGMA error logging, ALL_CAPS constant naming, display_priority column index,
  tooltip config old* guards, time milestone uses convert_time_to_text for natural format

version bump to 1.9.8

+++ MediaManager/src/QtClasses/UI/StatsDialog.h
@@ -90,6 +90,6 @@ private:
     QVector<QPair<QDate, double>> m_heatmapCache;
     int m_cachedVideosToday = -1;
     double m_cachedWatchedTodaySec = -1.0;
-    int m_cachedDailyVideoGoal = -1;
+    double m_cachedDailyVideoGoal = -1.0;
     int m_cachedDailyTimeGoalSec = -1;
 };
